# sdks/sdkwork-models-backend-sdk/openapi/sdkwork-models-backend-api.openapi.json
{
  "openapi": "3.1.2",
  "info": {
    "title": "SDKWork Models Backend API",
    "version": "0.1.0",
    "description": "Intelligence catalog API owned by sdkwork-models. Materialized from composed host catalog mount OpenAPI authority."
  },
  "servers": [
    {
      "url": "/backend/v3/api"
    }
  ],
  "paths": {
...
    "/backend/v3/api/ai/video_profiles": {
      "get": {
        "description": "List catalog video generation profiles with optional vendor, mode, duration, and resolution filters.",
        "operationId": "videoProfiles.list",
        "parameters": [
          {
            "in": "query",
            "name": "vendor_code",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Filter by vendor code."
          },
          {
            "in": "query",
            "name": "region_code",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Filter by vendor region code."
          },
          {
            "in": "query",
            "name": "catalog_key",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Filter by model catalog key vendor/modelId."
          },
          {
            "in": "query",
            "name": "model_id",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Filter by model id when combined with vendor_code."
          },
          {
            "in": "query",
            "name": "generation_mode",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Filter by generation mode (text_to_video, image_to_video, ...)."
          },
          {
            "in": "query",
            "name": "duration_tier_code",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Filter by canonical duration tier (dur_5s, dur_10s, ...)."
          },
          {
            "in": "query",
            "name": "resolution",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Filter by output resolution (720p, 1080p, ...)."
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelRankingsListResult"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unauthorized"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Server Error"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Error response."
          }
        },
        "security": [
          {
            "AccessToken": [],
            "AuthToken": []
          }
        ],
        "summary": "List video generation profiles",
        "tags": [
          "ai"
        ],
        "x-contract-kind": "read",
        "x-file-targets": [],
        "x-read-sources": [
          "ai"
        ],
        "x-route-scope": "admin",
        "x-sdk-domain": "intelligence",
        "x-sdkwork-api-authority": "sdkwork-models-backend-api",
        "x-sdkwork-domain": "intelligence",
        "x-sdkwork-owner": "sdkwork-models",
        "x-sdkwork-resource": "videoProfiles",
        "x-sdkwork-source-route-crate": "sdkwork-routes-models-catalog-backend-api",
        "x-source-file": "tools/bootstrap_frontend_contract_from_route_manifest.py",
        "x-write-tables": []
      }
    },
    "/backend/v3/api/ai/models/{modelId}/video_profiles": {
      "get": {
        "description": "List video generation profiles bound to a model catalog key.",
        "operationId": "modelVideoProfiles.list",
        "parameters": [
          {
            "in": "path",
            "name": "modelId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Model id or catalog key vendor/modelId."
          },
          {
            "in": "query",
            "name": "vendor_code",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Required when model id is not a catalog key."
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelRankingsListResult"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unauthorized"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Server Error"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Error response."
          }
        },
        "security": [
          {
            "AccessToken": [],
            "AuthToken": []
          }
        ],
        "summary": "List model video generation profiles",
        "tags": [
          "ai"
        ],
        "x-contract-kind": "read",
        "x-file-targets": [],
        "x-read-sources": [
          "ai"
        ],
        "x-route-scope": "admin",
        "x-sdk-domain": "intelligence",
        "x-sdkwork-api-authority": "sdkwork-models-backend-api",
        "x-sdkwork-domain": "intelligence",
        "x-sdkwork-owner": "sdkwork-models",
        "x-sdkwork-resource": "modelVideoProfiles",
        "x-sdkwork-source-route-crate": "sdkwork-routes-models-catalog-backend-api",
        "x-source-file": "tools/bootstrap_frontend_contract_from_route_manifest.py",
        "x-write-tables": []
      }
    }
  },
  "components": {
    "securitySchemes": {
      "AccessToken": {
        "in": "header",
        "name": "Access-Token",
        "type": "apiKey"
      },
      "AuthToken": {
        "bearerFormat": "SDKWork auth token",
        "scheme": "bearer",
        "type": "http"
      }
    },
    "schemas": {
...
      "FieldError": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "field",
          "message"
        ],
        "properties": {
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "code": {
            "type": "integer",
            "format": "int32",
            "minimum": 40011,
            "maximum": 40099
          },
          "i18nKey": {
            "type": "string"
          },
          "params": {
            "type": "object",
            "additionalProperties": {
              "oneOf": [
                {
                  "type": "string"
                },
                {
                  "type": "number"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "boolean"
                }
              ]
            }
          }
        }
      },
      "JsonNull": {
        "description": "JSON null value.",
        "type": "null"
      },
      "JsonObject": {
        "additionalProperties": {
          "allOf": [
            {
              "$ref": "#/components/schemas/JsonValue"
            }
          ],
          "description": "Schema fragment used by json object."
        },
        "description": "JSON object with typed JSON values.",
        "type": "object"
      },
      "JsonValue": {
        "description": "JSON value accepted by flexible Claw Router metadata and extension maps.",
        "oneOf": [
          {
            "description": "String JSON value.",
            "type": "string"
          },
          {
            "description": "Number JSON value.",
            "type": "number"
          },
          {
            "description": "Integer JSON value.",
            "type": "integer"
          },
          {
            "description": "Boolean JSON value.",
            "type": "boolean"
          },
          {
            "description": "Array JSON value.",
            "items": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "type": "array"
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/JsonObject"
              }
            ],
            "description": "Object JSON value."
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/JsonNull"
              }
            ],
            "description": "Null JSON value."
          }
        ]
      },
...
      "ModelRankingsListResult": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SdkWorkApiResponse"
          },
          {
            "additionalProperties": false,
            "description": "Object schema used by model rankings list result.",
            "properties": {
              "data": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/ModelRankingsPage"
                  }
                ],
                "description": "Data field on model rankings list result."
              }
            },
            "required": [
              "data"
            ],
            "type": "object"
          }
        ],
        "description": "Model rankings list result schema exposed by Claw Router.",
        "x-operation-id": "modelRankings.list"
      },
      "ModelRankingsPage": {
        "additionalProperties": false,
        "description": "Model rankings page schema exposed by Claw Router.",
        "name": "ModelRankingsPage",
        "properties": {
          "history": {
            "description": "History field on model rankings page.",
            "items": {
              "additionalProperties": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/JsonValue"
                  }
                ],
                "description": "History field on model rankings page."
              },
              "description": "History field on model rankings page.",
              "type": "object"
            },
            "type": "array"
          },
          "items": {
            "description": "Items field on model rankings page.",
            "items": {
              "additionalProperties": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/JsonValue"
                  }
                ],
                "description": "Items field on model rankings page."
              },
              "description": "Items field on model rankings page.",
              "type": "object"
            },
            "type": "array"
          },
          "pageInfo": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PageInfo"
              }
            ],
            "description": "Page info field on model rankings page."
          },
          "source": {
            "additionalProperties": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/JsonValue"
                }
              ],
              "description": "Source field on model rankings page."
            },
            "description": "Source field on model rankings page.",
            "type": "object"
          }
        },
        "required": [
          "items",
          "history",
          "source",
          "pageInfo"
        ],
        "type": "object"
      },
...
      "PageInfo": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "mode"
        ],
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "offset",
              "cursor"
            ]
          },
          "page": {
            "type": "integer",
            "minimum": 1
          },
          "pageSize": {
            "type": "integer",
            "minimum": 1,
            "maximum": 200
          },
          "totalItems": {
            "type": "string",
            "pattern": "^[0-9]+$"
          },
          "totalPages": {
            "type": "integer",
            "minimum": 0
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ]
          },
          "hasMore": {
            "type": "boolean"
          }
        }
      },
      "ProblemDetail": {
        "type": "object",
        "additionalProperties": true,
        "required": [
          "type",
          "title",
          "status",
          "code",
          "traceId"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri-reference"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": 100,
            "maximum": 599
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "code": {
            "$ref": "#/components/schemas/SdkWorkPlatformErrorCode"
          },
          "traceId": {
            "type": "string",
            "format": "uuid",
            "description": "Server-owned request correlation id."
          },
          "i18nKey": {
            "type": "string",
            "description": "Optional stable localization key such as errors.result.40001."
          },
          "locale": {
            "type": "string",
            "description": "Optional effective BCP 47 locale used by framework message mapping."
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          }
        }
      },
      "SdkWorkApiResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "data",
          "traceId"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "enum": [
              0
            ],
            "default": 0,
            "minimum": 0,
            "maximum": 0
          },
          "data": {
            "description": "Operation-specific payload typed per response schema."
          },
          "traceId": {
            "type": "string",
            "format": "uuid",
            "description": "Server-owned request correlation id."
          }
        }
      },
      "SdkWorkPlatformErrorCode": {
        "type": "integer",
        "format": "int32",
        "minimum": 40001,
        "maximum": 79999,
        "description": "Platform or domain error code per API_SPEC.md §15.3."
      },
...
    }
  }
}